fix: test and lint fixes
